PUT fixed + refactor
