+++ src/api/handlers/usersIdDELETE.js
@@ -1,0 +1,7 @@
+const usersIdDELETE = (_, res, collection, uuid) => {
+  collection.deleteUser(uuid);
+  res.statusCode = 204;
+  res.end();
+};
+
+export default usersIdDELETE;
